Correct references from 'Nugram' to 'Nutgram' across docs
This commit is addressing the error within documentation, where the class is erroneously referred to as 'Nugram'. All instances have been changed to the correct 'Nutgram'. This is reflected in the upgrading and usage docs across all the versions. This change will avoid confusing users who are following the documentation for implementation.